Trunk DP IL: Added int casting to 170_DangerousAllure
git-svn-id: http://svn.l2jfree.com/l2jfree@1877 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/L2_DataPack_IL/data/jscript/quests/170_DangerousAllure/__init__.py
@@ -45,7 +45,7 @@ class Quest (JQuest) :
    elif cond :
       if st.getQuestItemsCount(NIGHTMARE_CRYSTAL) :
          htmltext = "30305-06.htm"
-         st.giveItems(57,102680 * Config.RATE_QUESTS_REWARD)
+         st.giveItems(57, (int)(102680 * Config.RATE_QUESTS_REWARD))
          st.takeItems(NIGHTMARE_CRYSTAL,-1)
          st.setState(COMPLETED)
          st.playSound("ItemSound.quest_finish")
